GH-2584: Moved constant to the occ folder.
